Fix si4010_xram_write() for writes > 0xff bytes
The 'len' argument was only 8-bit, making writes bigger then 0xff bytes
impossible.

+++ src/si4010.c
@@ -423,7 +423,7 @@ int si4010_xram_read(uint16_t addr, size_t len, void *buf)
 	return 0;
 }
 
-int si4010_xram_write(uint16_t addr, uint8_t len, const void *buf)
+int si4010_xram_write(uint16_t addr, size_t len, const void *buf)
 {
 	uint8_t *bbuf = (uint8_t *) buf;
 	size_t blen;
